Switch GroupedCollection from struct to trait.
Instead of wrapping HashMap, GroupedCollection is now a trait that
HashMap and BTreeMap both implement and that other collections may
implement in the future as well. It is meant as an abstraction layer so
that code designed around GroupedCollection can swap out the data
structure underneath without affecting the surrounding code.