Add run_command function. WIP.
The run_command function provides a testable interface for running an
external command the way the groupby binary needs. It's part of the
ongoing push to provide an easy interface for parallel execution of
commands when the -c option is given.

I've run out of time for the day, so I'm committing what's done:
everything from std::process::Command up to and including spawn().
The groupby binary does not yet use this code, as it's not complete. The
binary remains unchanged and working.