Implement most of top-level output_results; WIP.
This commit focuses on abstracting the running of commands in a sensible
way that's intended to decouple deeper layers of logic from
groupby::command_line::options and any other top-level types, provide a
flexible interface for benchmarking different data structures and
execution strategies, and so on.

To the best of my knowledge, the code in this commit is correct, but:

- I haven't written tests yet (so assume it's all wrong!)
- It doesn't yet print the results
- The groupby binary doesn't yet use it

Follow-up will most likely come tomorrow.