Refactor output results (#3)
* Extract method Separator.sep().

Returns a static str corresponding to the enum variant.

* Replace match statement with unwrap_or_else.

* Add CommandReport.

* Add run_command function. WIP.

The run_command function provides a testable interface for running an
external command the way the groupby binary needs. It's part of the
ongoing push to provide an easy interface for parallel execution of
commands when the -c option is given.

I've run out of time for the day, so I'm committing what's done:
everything from std::process::Command up to and including spawn().
The groupby binary does not yet use this code, as it's not complete. The
binary remains unchanged and working.

* Add temporary refactoring comments.

These comments are in all caps to easily distinguish them from the
existing comments on the code being refactored. They lay out the overall
plan for refactoring and will be replaced once the code under
development is ready to roll out.

* Remove variable declaration; return it directly.

I now realize it makes sense to break the existing functionality of
groupby into multiple functions and to let this be the end of the
command function. THerefore, there's no need for a variable.

* Add reader/writer getters for child objects.

They're not tested, yet; that comes next.

* Add comments for private traits.

* Changes trait CommandReport to public.

* Reorganize run_command types; implement stdin.

* Replace stdout with wait_with_output.

* Rework CommandReport for vector output.

Also fix issues with mutability and sharing by creating
CommandReportInteriorMutable.

* Change CommandReport to accept any report type.

Originally, it was designed around the assumption that it would read a
BufReader to glean output in the form of a Vec<u8>; now that this
assumption is lifted, CommandReport can become fully general.

* Fix deadlock in CommandHandle::wait_with_output.

This commit resolves the deadlock that occurs when the user has
taken Child::stdin and then calls wait_with_output without manually
dropping the taken stdin value. By moving the StandardInput value from
CommandHandle::stdin() to CommandHandle::stdin, we allow
CommandHandle::wait_with_output() to (a) drop stdin itself to prevent
this deadlock and (b) fail to compile if the user has moved stdin,
thereby once again preventing this deadlock.

* Re-export CommandReportInteriorMutable.

* Use new command-running library functions.

This is not the final stage, but it's a fully working, single-threaded
prototype, ready for fine tuning and multithreading.

* Add TODO for removing Arc from command_report.

I don't think it's actually needed, but I don't want to remove it until
I verify that I'm correct. AFAICT a parallel iterator should not need
Arc, and since immutable references are Copy, I think each thread should
be able to share such a reference just fine without Arc. Furthermore, it
seems to me that implementing CommandReportInteriorMutable for Mutex<CR>
(and maybe renaming it to CommandReportMutex) should work just fine for
Arc<Mutex<CR>>, too.

* Sort dependencies by name.

* Switch output_results to use Rayon par_iter.

On my development box, a test sequence I've been running drops from ~560
ms to ~140ms with this commit.

* Reorganize run_command module.

Remove duplicated Command prefixes in names.
Fix Clippy issue: mod.rs and run_command.rs: rename to run.rs.
Try to make names as intuitive and clear as possible.

* Implement reporting for Mutex<Report>: remove Arc.

Previously reporting was only implemented for Arc<Mutex<Report>>, but
this was unnecessarily restrictive. Arc<Mutex<Report>> still works, but
so does Mutex<Report>. If needed, it should be simple to implement other
interior mutability techniques for ReportInteriorMutable, but at this
time, I see no reason to do so.

* Fix broken documentation link from refactoring.

* Rename StandardInput, methods to RecordWriter.

* Document rationale for not piping stderr in run().

* Implement most of top-level output_results; WIP.

This commit focuses on abstracting the running of commands in a sensible
way that's intended to decouple deeper layers of logic from
groupby::command_line::options and any other top-level types, provide a
flexible interface for benchmarking different data structures and
execution strategies, and so on.

To the best of my knowledge, the code in this commit is correct, but:

- I haven't written tests yet (so assume it's all wrong!)
- It doesn't yet print the results
- The groupby binary doesn't yet use it

Follow-up will most likely come tomorrow.

* Improve RecordWriter tests.

- Add MockWriter struct and use it to verify behavior.
- Clean up tests in various ways.

* Add unit tests.

* Add writing command results.

* Add writing results when not running a command.

I think output_results may finally be complete and ready to integrate
into groupby, but that decision comes tomorrow.

* Use new output_results. bin/groupby is refactored!

There's still work to do before merging the branch, including giving
additional consideration to tons of details and completing
documentation, but the binary is now fully refactored and hopefully very
easy to understand. Plus, it's all under test!

* Add doc comments for most of output_results.

More to come.

* Add doc comments for remaining functions in file.

* Rename run() to capture_command_output().

* Add module doc comment.

* Add example for output_results.

* Change ShellCommandsOptions fields to public.

They should always have been public, but I didn't realize the oversight
until I wrote a doctest that used the struct. Whoops!

* Add doctest for capture_command_output.

* Add module doc for command_line::run_command.

* Document command_line::run_command:: child ::*;

* Document command_line::run_command::command::*.

* Document command_line::run_command::handle::*.

* Document command_line::run_command::record_writer.

* Document command_line::run_command::mock_child::*.

* Document command_line::run_command::mock_command.

* Change report to be generic over key types.

The initial restriction was to ensure smooth development, but when
writing documentation and example code, it became clear that this was an
unnecessary restriction. The code in this library does not currently
need such generality, but it was simple to add and free.

* Document command_line::run_command::report better.

* Add example for ReportInteriorMutable.

* Document command_line::run_command::run::*.