Extract write_results() method for DRYness.
The refactored code was right on the threshold of needing refactoring
for DRYness, but it didn't initially make sense to do so, because the
logic for each of the cases seemed subtly different. Now, it's time to
refactor out the duplication in preparation for adding a new
command-line option that will change the formatting of the final report.