Reorganize run_command and output_results.
The intents here are to (a) improve physical organization of code and
(b) expose the clearest possible public API in the command_line module.

Major changes:
- Rename run_command module to command_runner. IMO this is a much more
appropriate name, for numerous reasons. It also resolves the name
conflict with the run_command function.

- Split output_results.rs into run_command.rs, write_results.rs, and
test_helpers.rs. This is the logical final outcome of splitting
output_results() into run_command() and write_results().