Add capture group option (#6)
Also rewrite `crate::matchers::string::match_regex` for improved clarity
and obviousness.

Rename following methods:

- grouping_options        -> groupers
- grouping_heading        -> groupers_heading
- group_by_first_chars    -> groupers_by_first_chars
- group_by_last_chars     -> groupers_by_last_chars
- group_by_regex          -> groupers_by_regex
- group_by_file_extension -> groupers_by_file_extension
- group_by_counter        -> groupers_by_counter

There were two reasons for this change:

1. The method names, while they were easy to read, lacked consistency.
   They were inconsistent with both each other and the other options in
   args. While groupers_by_first_chars is slightly less readable than
   group_by_first_chars, it is explicitly clear that it belongs inside
   groupers. This organization is increasingly important as the
   available options continue to expand.

2. We're adding a new section for setting grouper options, i.e. options
   that change the behavior of the chosen grouper. This creates a clear
   name conflict between, e.g., grouping_options for choosing a grouper
   and grouper_options for customizing the chosen grouper. This commit
   resolve that confusion unambiguously.

The justification for keeping the "_by" in the methods that add specific
groupers (e.g. groupers_by_first_chars) is that this treats "_by" as
part of the individual grouper's name, rather than a strict convention.
This allows for the possibility that future groupers might not use
"_by", if it's inappropriate, hopefully without creating a confusing
naming scheme in args as a result.